import correctections from renamed genericConfigFacade
