+++ gameModel/libs/ClientScript/game/common/impacts/implementation/emptyImpact.ts
@@ -1,8 +1,8 @@
-import { IDescriptor, Indexed, Typed } from "../../interfaces"
+import { IDescriptor, Indexed, Typed } from '../../interfaces';
 
 /**
  * Empty impact that does nothing, used for scenarist interface purposes
  */
 export interface EmptyImpact extends IDescriptor, Typed, Indexed {
-  type : 'empty'
+  type: 'empty';
 }
